+++ tests/templates.py
@@ -2576,3 +2576,99 @@ the full function diagram is emitted.
 
 
 """
+
+
+@templates.register
+class _TestCommitPolicy(BaseTemplate):
+    @property
+    def template(self) -> str:
+        return """
+policies:
+  - type: commit
+    spec:
+      header:
+        length: 72
+        imperative: true
+        invalidLastCharacters: .
+      dco: true
+      gpg:
+        required: true
+      spellcheck:
+        locale: US
+      conventional:
+        types:
+          - add
+          - build
+          - bump
+          - change
+          - chore
+          - ci
+          - doc
+          - fix
+          - hack
+          - refactor
+          - security
+          - style
+          - test
+          - remove
+        scopes:
+          - deps
+          - deps-dev
+        descriptionLength: 72
+"""
+
+    @property
+    def expected(self) -> str:
+        return """\
+<!--
+This file is auto-generated and any changes made to it will be overwritten
+-->
+# Commit Policy
+
+## Header
+
+Length: 72
+
+Imperative: True
+
+Invalid last characters: .
+
+## Dco
+
+True
+
+## Gpg
+
+Required: True
+
+## Spellcheck
+
+Locale: US
+
+## Conventional
+
+### Types
+
+- add
+- build
+- bump
+- change
+- chore
+- ci
+- doc
+- fix
+- hack
+- refactor
+- security
+- style
+- test
+- remove
+
+### Scopes
+
+- deps
+- deps-dev
+
+Description length: 72
+
+"""