+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/pyaud-plugins/compare/v0.10.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Add `CommitPolicy` plugin
 - Add `AboutTests` plugin
 
 [0.10.0](https://github.com/jshwi/pyaud-plugins/releases/tag/v0.10.0) - 2023-01-04